refactor: move message modules and update related imports
