+++ src/include/domains/access/handlers.py
@@ -11,7 +11,7 @@ from include.database.models.identity import (
 )
 from include.database.session import Session
 from include.domains.access.permissions import Permissions
-from include.domains.operations.messages import Messages as smsg
+from include.messages import Messages as smsg
 from include.transport.connection import ConnectionHandler
 from include.transport.request_handler import RequestHandler
