+++ src/include/domains/keyrings/handlers/keyrings.py
@@ -27,7 +27,7 @@ from include.database.models.identity import User
 from include.database.models.keyrings import UserKey
 from include.database.session import Session
 from include.domains.access.permissions import Permissions
-from include.domains.operations.messages import Messages as smsg
+from include.messages import Messages as smsg
 from include.transport.connection import ConnectionHandler
 from include.transport.request_handler import RequestHandler
